Renamed MLPlanProprietaryPropertyTester
